+++ SandboxiePlus/SandMan/Windows/OptionsGrouping.cpp
@@ -30,8 +30,10 @@ void COptionsWindow::LoadGroups()
 				continue;
 			QTreeWidgetItem* pSubItem = new QTreeWidgetItem();
 			SetProgramItem(Entries[i], pSubItem, 0);
+			pSubItem->setFlags(pSubItem->flags() | Qt::ItemIsEditable);
 			pItem->addChild(pSubItem);
 		}
+		pItem->setFlags(pItem->flags() | Qt::ItemIsEditable);
 		ui.treeGroups->addTopLevelItem(pItem);
 	}
 	
@@ -149,52 +151,26 @@ void COptionsWindow::OnAddGroup()
 	QTreeWidgetItem* pItem = new QTreeWidgetItem();
 	pItem->setText(0, Value);
 	pItem->setData(0, Qt::UserRole, "<" + Value + ">");
+	pItem->setFlags(pItem->flags() | Qt::ItemIsEditable);
 	ui.treeGroups->addTopLevelItem(pItem);
 
 	m_GroupsChanged = true;
 	OnOptChanged();
 }
 
-void COptionsWindow::AddProgToGroup(QTreeWidget* pTree, const QString& Groupe)
+void COptionsWindow::AddProgToGroup(QTreeWidget* pTree, const QString& Groupe, bool disabled)
 {
 	QString Value = SelectProgram();
 	if (Value.isEmpty())
 		return;
 
 	QTreeWidgetItem* pItem = new QTreeWidgetItem();
+	pItem->setCheckState(0, disabled ? Qt::Unchecked : Qt::Checked);
 	SetProgramItem(Value, pItem, 0);
+	pItem->setFlags(pItem->flags() | Qt::ItemIsEditable);
 	pTree->addTopLevelItem(pItem);
 
-	AddProgToGroup(Value, Groupe);
-}
-
-void COptionsWindow::AddProgToGroup(const QString& Value, const QString& Groupe)
-{
-	QTreeWidgetItem* pGroupItem = NULL;
-	for (int i = 0; i < ui.treeGroups->topLevelItemCount(); i++)
-	{
-		QTreeWidgetItem* pCurItem = ui.treeGroups->topLevelItem(i);
-		if (pCurItem->data(0, Qt::UserRole).toString().compare(Groupe, Qt::CaseInsensitive) == 0)
-		{
-			pGroupItem = pCurItem;
-			break;
-		}
-	}
-
-	if (!pGroupItem)
-	{
-		pGroupItem = new QTreeWidgetItem();
-		pGroupItem->setText(0, Groupe.mid(1, Groupe.length()-2));
-		pGroupItem->setData(0, Qt::UserRole, Groupe);
-		ui.treeGroups->addTopLevelItem(pGroupItem);
-	}
-
-	QTreeWidgetItem* pProgItem = new QTreeWidgetItem();
-	SetProgramItem(Value, pProgItem, 0);
-	pGroupItem->addChild(pProgItem);
-
-	m_GroupsChanged = true;
-	OnOptChanged();
+	AddProgramToGroup(Value, Groupe);
 }
 
 void COptionsWindow::DelProgFromGroup(QTreeWidget* pTree, const QString& Groupe)
@@ -235,26 +211,31 @@ void COptionsWindow::AddProgramToGroup(const QString& Program, const QString& Gr
 
 	QTreeWidgetItem* pSubItem = new QTreeWidgetItem();
 	SetProgramItem(Program, pSubItem, 0);
+	pSubItem->setFlags(pSubItem->flags() | Qt::ItemIsEditable);
 	pItem->addChild(pSubItem);
 
 	m_GroupsChanged = true;
 	OnOptChanged();
 }
 
-void COptionsWindow::DelProgramFromGroup(const QString& Program, const QString& Group)
+bool COptionsWindow::DelProgramFromGroup(const QString& Program, const QString& Group)
 {
 	QTreeWidgetItem* pItem = FindGroupByName(Group, true);
 
+	bool bFound = false;
 	for (int j = 0; j < pItem->childCount(); j++){
 		QTreeWidgetItem* pProgItem = pItem->child(j);
 		if (pProgItem->data(0, Qt::UserRole).toString().compare(Program, Qt::CaseInsensitive) == 0)  {
 			delete pProgItem;
+			bFound = true;
 			break;
 		}
 	}
 
 	m_GroupsChanged = true;
 	OnOptChanged();
+
+	return bFound;
 }
 
 QTreeWidgetItem* COptionsWindow::FindGroupByName(const QString& Group, bool bAdd)
@@ -273,6 +254,7 @@ QTreeWidgetItem* COptionsWindow::FindGroupByName(const QString& Group, bool bAdd
 		if (GroupName.length() > 2)
 			GroupName = GroupName.mid(1, GroupName.length() - 2);
 		pItem->setText(0, GroupName);
+		pItem->setFlags(pItem->flags() | Qt::ItemIsEditable);
 		ui.treeGroups->addTopLevelItem(pItem);
 		return pItem;
 	}
@@ -303,6 +285,7 @@ void COptionsWindow::OnAddProg()
 
 	QTreeWidgetItem* pSubItem = new QTreeWidgetItem();
 	SetProgramItem(Value, pSubItem, 0);
+	pSubItem->setFlags(pSubItem->flags() | Qt::ItemIsEditable);
 	pItem->addChild(pSubItem);
 
 	m_GroupsChanged = true;
@@ -326,9 +309,8 @@ void COptionsWindow::OnDelProg()
 	OnOptChanged();
 }
 
-void COptionsWindow::CopyGroupToList(const QString& Groupe, QTreeWidget* pTree)
+void COptionsWindow::CopyGroupToList(const QString& Groupe, QTreeWidget* pTree, bool disabled)
 {
-	pTree->clear();
 	for (int i = 0; i < ui.treeGroups->topLevelItemCount(); i++) 
 	{
 		QTreeWidgetItem* pItem = ui.treeGroups->topLevelItem(i);
@@ -339,7 +321,9 @@ void COptionsWindow::CopyGroupToList(const QString& Groupe, QTreeWidget* pTree)
 				QString Value = pItem->child(j)->data(0, Qt::UserRole).toString();
 
 				QTreeWidgetItem* pSubItem = new QTreeWidgetItem();
+				pSubItem->setCheckState(0, disabled ? Qt::Unchecked : Qt::Checked);
 				SetProgramItem(Value, pSubItem, 0);
+				pSubItem->setFlags(pSubItem->flags() | Qt::ItemIsEditable);
 				pTree->addTopLevelItem(pSubItem);
 			}
 			break;